Allow both static and dynamic linking with raylib
This commit includes raylib as a submodule for Mason, which by default
will be built when running make and the object files from raylib and
Mason will be archived into a single static library. If the user instead
wishes to link with raylib dynamically, the environment variables
RAYLIB_LIBTYPE and RAYLIB_INCLUDE_DIR must be set to DYNAMIC and to the
path which includes the headers for raylib respectivelly, and the raylib
submodule does not need to be cloned.